applications/backend/pyproject.toml
[tool.poetry]
name = "wordapp-backend"
version = "0.1.0"
description = "The experimental project of making an English-learning application mainly for the internal usage"
authors = ["Yoshiki Kubotani"]
maintainers = [
    "yoshiki kubotani <yoshikikubotani.lab@gmail.com>",
]
license = "MIT"
readme = "README.md"
keywords = [
    "wordapp",
    "Python",
    "FastAPI",
    "Education",
]
packages = [{include = "wordapp_backend"}]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Framework :: FastAPI",
    "Framework :: Flake8",
    "Framework :: Pydantic",
    "Framework :: Pytest",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Natural Language :: Japanese",
    "Programming Language :: Python :: 3.10",
    "Topic :: Education",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Version Control :: Git",
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.0"
gunicorn = "^21.2.0"
uvicorn = "^0.23.2"
orjson = "^3.9.5"
psycopg2-binary = "^2.9.7"
pydantic = "^2.3.0"
networkx = "^3.1"
websockets = "^11.0.3"
redis = "^5.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.286"
flake8 = "^6.0.0"
mypy = "^1.5.1"
pytest-cov = "^4.1.0"

[tool.black]
line-length = 88

[tool.mypy]
python_version = "3.10"
# following setting is same as pysen
# https://github.com/pfnet/pysen/blob/main/setup.cfg#L12
check_untyped_defs = true
disallow_any_decorated = false
disallow_any_generics = false
disallow_any_unimported = false
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
disallow_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.ruff]
select = ["E", "F", "B", "I", "N", "D"]
ignore = ["E501", "B006", "D100", "D104", "D205"]
line-length = 88
target-version = "py38"
fixable = ["E", "F", "I", "D"]
unfixable = ["F401", "F841"]

[tool.ruff.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
